Swing component in develop

+++ src/main/java/sofware/ulpgc/project/model/Money.java
@@ -1,3 +1,3 @@
 package sofware.ulpgc.project.model;
 
-public record Money(int amount, Currency currency){}
+public record Money(double amount, Currency currency){}
